split signup/login js
