--- movier.js
@@ -1,5 +1,0 @@
-const movier = require('movier');
-
-movier.searchTitleByName('cop out').then(element => {
-    console.log(element[0]);
-})
