+++ views/signup.handlebars
@@ -1,4 +1,3 @@
-<link rel="stylesheet" href="/stylesheets/style.css">
 <form class="signup-form">
   <div>
     <label for="username-signup">Username:</label>
@@ -16,8 +15,8 @@
     <button type="submit" class="sgnupsbmt">Sign up</button>
   </div>
   <div>
-    <a href="/login"><button type="submit" class="lginbtn">Login</button></a>
-  </div>
+    <a href="/login"><button type="submit" class="lginbtn">Login</button></a>  
+    </div>
 </form>
 
-<script src="/js/login.js"></script>
+<script src="/js/signup.js"></script>
